Add parseFile to store values from text in struct
The readFile function is renamed to parseFile and is a function that
takes in a variable of type struct passenger and stores data fields
from a text file into the members of the struct. The file is read using
fopen and the file is read line by line using fgets. The string that
holds one line is tokenized into substrings using strtok and each
substring represents a data field that is stored in the struct that is
passed into the function.

Parsing files is necessary to generate a priority queue based on
passenger information. The parseFile function allows the data to be read
into the code and stored in a struct for easy access.