Add priority numbers creation and print passenger
Add genPriorityNum to generate a priority number for every passenger
in the struct that is passed into the function. Add printPassData
to print the values in each entry in the passenger struct that is passed
into the function.

Generating priority numbers is necessary for storing the passengers
in a max heap and printing the passengers allows for easy display
of the passenger data.