Add insertHeap to make maxHeap from passenger list
The struct maxHeap was added to create a maxHeap from passenger data.
The function heapify was added to be used in the function insertHeap
to properly order the elements in the heap. The function insertHeap
takes in a heap and a passenger array and inserts the values from the
passenger array into the heap.

Inserting elements into a maxHeap is necessary for the passengers to be
ordered according to their priority number.