Add heapSort to sort passengers by high priority
The heapSort function was added to sort a passenger array by priority
number in decreasing order. The array is put into a minimum binary heap
and the passengers are reordered so that the passengers with the highest
priority number are first.

Adding heapSort creates the priority queue for the passengers.